feat: allow non-cli-connected execution
`noti-cli.json` should have an empty string as the `cli` key, or not
have it at all, in which case the first parameter to `noti-cli` will be
used as the cli to start.

Could be useful for default non-generator behaviour (e.g. watch for
end-stream and such, useful for any long-running operation)